+++ squidworld/src/main/java/com/github/yellowstonegames/world/TilingWorldMap.java
@@ -14,21 +14,21 @@ import com.github.yellowstonegames.grid.Noise;
 public class TilingWorldMap extends WorldMapGenerator {
     private static final long serialVersionUID = 1L;
     protected static final float terrainFreq = 0.95f, terrainRidgedFreq = 2.6f, heatFreq = 2.1f, moistureFreq = 2.125f, otherFreq = 3.375f;
-    private double minHeat0 = Double.POSITIVE_INFINITY, maxHeat0 = Double.NEGATIVE_INFINITY,
-            minHeat1 = Double.POSITIVE_INFINITY, maxHeat1 = Double.NEGATIVE_INFINITY,
-            minWet0 = Double.POSITIVE_INFINITY, maxWet0 = Double.NEGATIVE_INFINITY;
+    private float minHeat0 = Float.POSITIVE_INFINITY, maxHeat0 = Float.NEGATIVE_INFINITY,
+            minHeat1 = Float.POSITIVE_INFINITY, maxHeat1 = Float.NEGATIVE_INFINITY,
+            minWet0 = Float.POSITIVE_INFINITY, maxWet0 = Float.NEGATIVE_INFINITY;
 
     public final Noise terrain, terrainRidged, heat, moisture, otherRidged;
 
     /**
      * Constructs a concrete WorldMapGenerator for a map that can be used as a tiling, wrapping east-to-west as well
      * as north-to-south. Always makes a 256x256 map.
-     * Uses Noise as its noise generator, with 1.0 as the octave multiplier affecting detail.
-     * If you were using {@link TilingWorldMap#TilingWorldMap(long, int, int, Noise, double)}, then this would be the
-     * same as passing the parameters {@code 0x1337BABE1337D00DL, 256, 256, WorldMapGenerator.DEFAULT_NOISE, 1.0}.
+     * Uses Noise as its noise generator, with 1.0f as the octave multiplier affecting detail.
+     * If you were using {@link TilingWorldMap#TilingWorldMap(long, int, int, Noise, float)}, then this would be the
+     * same as passing the parameters {@code 0x1337BABE1337D00DL, 256, 256, WorldMapGenerator.DEFAULT_NOISE, 1.0f}.
      */
     public TilingWorldMap() {
-        this(0x1337BABE1337D00DL, 256, 256, WorldMapGenerator.DEFAULT_NOISE, 1.0);
+        this(0x1337BABE1337D00DL, 256, 256, WorldMapGenerator.DEFAULT_NOISE, 1.0f);
     }
 
     /**
@@ -37,13 +37,13 @@ public class TilingWorldMap extends WorldMapGenerator {
      * Takes only the width/height of the map. The initial seed is set to the same large long
      * every time, and it's likely that you would set the seed when you call {@link #generate(long, long)}. The width and
      * height of the map cannot be changed after the fact, but you can zoom in.
-     * Uses Noise as its noise generator, with 1.0 as the octave multiplier affecting detail.
+     * Uses Noise as its noise generator, with 1.0f as the octave multiplier affecting detail.
      *
      * @param mapWidth  the width of the map(s) to generate; cannot be changed later
      * @param mapHeight the height of the map(s) to generate; cannot be changed later
      */
     public TilingWorldMap(int mapWidth, int mapHeight) {
-        this(0x1337BABE1337D00DL, mapWidth, mapHeight, DEFAULT_NOISE, 1.0);
+        this(0x1337BABE1337D00DL, mapWidth, mapHeight, DEFAULT_NOISE, 1.0f);
     }
 
     /**
@@ -52,14 +52,14 @@ public class TilingWorldMap extends WorldMapGenerator {
      * Takes an initial seed and the width/height of the map. The {@code initialSeed}
      * parameter may or may not be used, since you can specify the seed to use when you call {@link #generate(long, long)}.
      * The width and height of the map cannot be changed after the fact, but you can zoom in.
-     * Uses Noise as its noise generator, with 1.0 as the octave multiplier affecting detail.
+     * Uses Noise as its noise generator, with 1.0f as the octave multiplier affecting detail.
      *
      * @param initialSeed the seed for the LaserRandom this uses; this may also be set per-call to generate
      * @param mapWidth    the width of the map(s) to generate; cannot be changed later
      * @param mapHeight   the height of the map(s) to generate; cannot be changed later
      */
     public TilingWorldMap(long initialSeed, int mapWidth, int mapHeight) {
-        this(initialSeed, mapWidth, mapHeight, DEFAULT_NOISE, 1.0);
+        this(initialSeed, mapWidth, mapHeight, DEFAULT_NOISE, 1.0f);
     }
 
     /**
@@ -69,9 +69,9 @@ public class TilingWorldMap extends WorldMapGenerator {
      * parameter may or may not be used, since you can specify the seed to use when you call
      * {@link #generate(long, long)}. The width and height of the map cannot be changed after the fact, but you can zoom
      * in. Any seed supplied to the Noise given to this (if it takes one) will be ignored, and
-     * {@link Noise#getNoiseWithSeed(double, double, double, double, long)} will be used to specify the seed many
+     * {@link Noise#getNoiseWithSeed(float, float, float, float, int)} will be used to specify the seed many
      * times. The detail level, which is the {@code octaveMultiplier} parameter that can be passed to another
-     * constructor, is always 1.0 with this constructor.
+     * constructor, is always 1.0f with this constructor.
      *
      * @param initialSeed    the seed for the LaserRandom this uses; this may also be set per-call to generate
      * @param mapWidth       the width of the map(s) to generate; cannot be changed later
@@ -79,61 +79,61 @@ public class TilingWorldMap extends WorldMapGenerator {
      * @param noiseGenerator an instance of a noise generator capable of 4D noise, recommended to be {@link Noise#instance}
      */
     public TilingWorldMap(long initialSeed, int mapWidth, int mapHeight, final Noise noiseGenerator) {
-        this(initialSeed, mapWidth, mapHeight, noiseGenerator, 1.0);
+        this(initialSeed, mapWidth, mapHeight, noiseGenerator, 1.0f);
     }
 
     /**
      * Constructs a concrete WorldMapGenerator for a map that can be used as a tiling, wrapping east-to-west as well
      * as north-to-south. Takes an initial seed, the width/height of the map, and parameters for noise
      * generation (a {@link Noise} implementation, which is usually {@link Noise#instance}, and a
-     * multiplier on how many octaves of noise to use, with 1.0 being normal (high) detail and higher multipliers
+     * multiplier on how many octaves of noise to use, with 1.0f being normal (high) detail and higher multipliers
      * producing even more detailed noise when zoomed-in). The {@code initialSeed} parameter may or may not be used,
      * since you can specify the seed to use when you call {@link #generate(long, long)}. The width and height of the map
-     * cannot be changed after the fact, but you can zoom in. Any seed supplied to the Noise given to this (if it takes one) will be ignored, and
-     * {@link Noise#getNoiseWithSeed(double, double, double, double, long)} will be used to specify the seed many
-     * times. The {@code octaveMultiplier} parameter should probably be no lower than 0.5, but can be arbitrarily
+     * cannot be changed after the fact, but you can zoom in. Any seed supplied to the Noise given to this will be
+     * ignored, and {@link Noise#getNoiseWithSeed(float, float, float, float, int)} will be used to specify the seed
+     * many times. The {@code octaveMultiplier} parameter should probably be no lower than 0.5f, but can be arbitrarily
      * high if you're willing to spend much more time on generating detail only noticeable at very high zoom;
-     * normally 1.0 is fine and may even be too high for maps that don't require zooming.
+     * normally 1.0f is fine and may even be too high for maps that don't require zooming.
      *
      * @param initialSeed      the seed for the LaserRandom this uses; this may also be set per-call to generate
      * @param mapWidth         the width of the map(s) to generate; cannot be changed later
      * @param mapHeight        the height of the map(s) to generate; cannot be changed later
      * @param noiseGenerator   an instance of a noise generator capable of 4D noise, almost always {@link Noise}
-     * @param octaveMultiplier used to adjust the level of detail, with 0.5 at the bare-minimum detail and 1.0 normal
+     * @param octaveMultiplier used to adjust the level of detail, with 0.5f at the bare-minimum detail and 1.0f normal
      */
-    public TilingWorldMap(long initialSeed, int mapWidth, int mapHeight, final Noise noiseGenerator, double octaveMultiplier) {
+    public TilingWorldMap(long initialSeed, int mapWidth, int mapHeight, final Noise noiseGenerator, float octaveMultiplier) {
         super(initialSeed, mapWidth, mapHeight);
         terrain = new Noise(noiseGenerator);
         terrain.setFrequency(terrain.getFrequency() * terrainFreq);
         terrain.setNoiseType(terrain.getNoiseType() | 1);
-        terrain.setFractalOctaves((int) (0.5 + octaveMultiplier * 8));
+        terrain.setFractalOctaves((int) (0.5f + octaveMultiplier * 8));
         terrain.setFractalLacunarity(0.5f);
         terrain.setFractalGain(2f);
 
         terrainRidged = new Noise(noiseGenerator);
         terrainRidged.setFrequency(terrainRidged.getFrequency() * terrainRidgedFreq);
         terrainRidged.setNoiseType(terrainRidged.getNoiseType() | 1);
-        terrainRidged.setFractalOctaves((int) (0.5 + octaveMultiplier * 10));
+        terrainRidged.setFractalOctaves((int) (0.5f + octaveMultiplier * 10));
         terrainRidged.setFractalType(Noise.RIDGED_MULTI);
 
         heat = new Noise(noiseGenerator);
         heat.setFrequency(heat.getFrequency() * heatFreq);
         heat.setNoiseType(heat.getNoiseType() | 1);
-        heat.setFractalOctaves((int) (0.5 + octaveMultiplier * 3));
+        heat.setFractalOctaves((int) (0.5f + octaveMultiplier * 3));
         heat.setFractalLacunarity(0.5f);
         heat.setFractalGain(2f);
 
         moisture = new Noise(noiseGenerator);
         moisture.setFrequency(moisture.getFrequency() * moistureFreq);
         moisture.setNoiseType(moisture.getNoiseType() | 1);
-        moisture.setFractalOctaves((int) (0.5 + octaveMultiplier * 4));
+        moisture.setFractalOctaves((int) (0.5f + octaveMultiplier * 4));
         moisture.setFractalLacunarity(0.5f);
         moisture.setFractalGain(2f);
 
         otherRidged = new Noise(noiseGenerator);
         otherRidged.setFrequency(otherRidged.getFrequency() * otherFreq);
         otherRidged.setNoiseType(otherRidged.getNoiseType() | 1);
-        otherRidged.setFractalOctaves((int) (0.5 + octaveMultiplier * 6));
+        otherRidged.setFractalOctaves((int) (0.5f + octaveMultiplier * 6));
         otherRidged.setFractalType(Noise.RIDGED_MULTI);
     }
 
@@ -159,21 +159,21 @@ public class TilingWorldMap extends WorldMapGenerator {
     }
 
     protected void regenerate(int startX, int startY, int usedWidth, int usedHeight,
-                              double landMod, double heatMod, long stateA, long stateB) {
+                              float landMod, float heatMod, long stateA, long stateB) {
         boolean fresh = false;
         if (cacheA != stateA || cacheB != stateB || landMod != landModifier || heatMod != heatModifier) {
-            minHeight = Double.POSITIVE_INFINITY;
-            maxHeight = Double.NEGATIVE_INFINITY;
-            minHeat0 = Double.POSITIVE_INFINITY;
-            maxHeat0 = Double.NEGATIVE_INFINITY;
-            minHeat1 = Double.POSITIVE_INFINITY;
-            maxHeat1 = Double.NEGATIVE_INFINITY;
-            minHeat = Double.POSITIVE_INFINITY;
-            maxHeat = Double.NEGATIVE_INFINITY;
-            minWet0 = Double.POSITIVE_INFINITY;
-            maxWet0 = Double.NEGATIVE_INFINITY;
-            minWet = Double.POSITIVE_INFINITY;
-            maxWet = Double.NEGATIVE_INFINITY;
+            minHeight = Float.POSITIVE_INFINITY;
+            maxHeight = Float.NEGATIVE_INFINITY;
+            minHeat0 = Float.POSITIVE_INFINITY;
+            maxHeat0 = Float.NEGATIVE_INFINITY;
+            minHeat1 = Float.POSITIVE_INFINITY;
+            maxHeat1 = Float.NEGATIVE_INFINITY;
+            minHeat = Float.POSITIVE_INFINITY;
+            maxHeat = Float.NEGATIVE_INFINITY;
+            minWet0 = Float.POSITIVE_INFINITY;
+            maxWet0 = Float.NEGATIVE_INFINITY;
+            minWet = Float.POSITIVE_INFINITY;
+            maxWet = Float.NEGATIVE_INFINITY;
             cacheA = stateA;
             cacheB = stateB;
             fresh = true;
@@ -182,16 +182,16 @@ public class TilingWorldMap extends WorldMapGenerator {
         long seedA = rng.nextLong(), seedB = rng.nextLong(), seedC = rng.nextLong();
         int t;
 
-        landModifier = (landMod <= 0) ? rng.nextDouble(0.1875) + 0.99 : landMod;
-        heatModifier = (heatMod <= 0) ? rng.nextDouble(0.45) * (rng.nextDouble() - 0.5) + 1.1 : heatMod;
+        landModifier = (landMod <= 0) ? rng.nextFloat(0.1875f) + 0.99f : landMod;
+        heatModifier = (heatMod <= 0) ? rng.nextFloat(0.45f) * (rng.nextFloat() - 0.5f) + 1.1f : heatMod;
 
-        double p, q,
+        float p, q,
                 ps, pc,
                 qs, qc,
                 h, temp,
-                i_w = 6.283185307179586 / width, i_h = 6.283185307179586 / height,
-                xPos = startX, yPos = startY, i_uw = usedWidth / (double) width, i_uh = usedHeight / (double) height;
-        double[] trigTable = new double[width << 1];
+                i_w = 6.283185307179586f / width, i_h = 6.283185307179586f / height,
+                xPos = startX, yPos = startY, i_uw = usedWidth / (float) width, i_uh = usedHeight / (float) height;
+        float[] trigTable = new float[width << 1];
         for (int x = 0; x < width; x++, xPos += i_uw) {
             p = xPos * i_w;
             trigTable[x << 1] = TrigTools.sin(p);
@@ -205,13 +205,13 @@ public class TilingWorldMap extends WorldMapGenerator {
                 ps = trigTable[xt++];//TrigTools.sin(p);
                 pc = trigTable[xt++];//TrigTools.cos(p);
                 heightData[x][y] = (h = terrain.getNoiseWithSeed(pc +
-                                terrainRidged.getNoiseWithSeed(pc, ps, qc, qs, seedB - seedA) * 0.25,
-                        ps, qc, qs, seedA) + landModifier - 1.0);
+                                terrainRidged.getNoiseWithSeed(pc, ps, qc, qs, seedB - seedA) * 0.25f,
+                        ps, qc, qs, seedA) + landModifier - 1.0f);
                 heatData[x][y] = (p = heat.getNoiseWithSeed(pc, ps, qc
-                                + 0.375 * otherRidged.getNoiseWithSeed(pc, ps, qc, qs, seedB + seedC)
+                                + 0.375f * otherRidged.getNoiseWithSeed(pc, ps, qc, qs, seedB + seedC)
                         , qs, seedB));
                 moistureData[x][y] = (temp = moisture.getNoiseWithSeed(pc, ps, qc, qs
-                                + 0.375 * otherRidged.getNoiseWithSeed(pc, ps, qc, qs, seedC + seedA)
+                                + 0.375f * otherRidged.getNoiseWithSeed(pc, ps, qc, qs, seedC + seedA)
                         , seedC));
                 minHeightActual = Math.min(minHeightActual, h);
                 maxHeightActual = Math.max(maxHeightActual, h);
@@ -231,44 +231,44 @@ public class TilingWorldMap extends WorldMapGenerator {
             maxHeightActual = Math.max(maxHeightActual, maxHeight);
 
         }
-        double
-                heatDiff = 0.8 / (maxHeat0 - minHeat0),
-                wetDiff = 1.0 / (maxWet0 - minWet0),
+        float
+                heatDiff = 0.8f / (maxHeat0 - minHeat0),
+                wetDiff = 1.0f / (maxWet0 - minWet0),
                 hMod,
-                halfHeight = (height - 1) * 0.5, i_half = 1.0 / halfHeight;
+                halfHeight = (height - 1) * 0.5f, i_half = 1.0f / halfHeight;
         yPos = startY;
-        ps = Double.POSITIVE_INFINITY;
-        pc = Double.NEGATIVE_INFINITY;
+        ps = Float.POSITIVE_INFINITY;
+        pc = Float.NEGATIVE_INFINITY;
 
         for (int y = 0; y < height; y++, yPos += i_uh) {
             temp = Math.abs(yPos - halfHeight) * i_half;
-            temp *= (2.4 - temp);
-            temp = 2.2 - temp;
+            temp *= (2.4f - temp);
+            temp = 2.2f - temp;
             for (int x = 0; x < width; x++) {
                 h = heightData[x][y];
                 heightCodeData[x][y] = (t = codeHeight(h));
-                hMod = 1.0;
+                hMod = 1.0f;
                 switch (t) {
                     case 0:
                     case 1:
                     case 2:
                     case 3:
-                        h = 0.4;
-                        hMod = 0.2;
+                        h = 0.4f;
+                        hMod = 0.2f;
                         break;
                     case 6:
-                        h = -0.1 * (h - forestLower - 0.08);
+                        h = -0.1f * (h - forestLower - 0.08f);
                         break;
                     case 7:
-                        h *= -0.25;
+                        h *= -0.25f;
                         break;
                     case 8:
-                        h *= -0.4;
+                        h *= -0.4f;
                         break;
                     default:
-                        h *= 0.05;
+                        h *= 0.05f;
                 }
-                heatData[x][y] = (h = (((heatData[x][y] - minHeat0) * heatDiff * hMod) + h + 0.6) * temp);
+                heatData[x][y] = (h = (((heatData[x][y] - minHeat0) * heatDiff * hMod) + h + 0.6f) * temp);
                 if (fresh) {
                     ps = Math.min(ps, h); //minHeat0
                     pc = Math.max(pc, h); //maxHeat0
@@ -280,10 +280,10 @@ public class TilingWorldMap extends WorldMapGenerator {
             maxHeat1 = pc;
         }
         heatDiff = heatModifier / (maxHeat1 - minHeat1);
-        qs = Double.POSITIVE_INFINITY;
-        qc = Double.NEGATIVE_INFINITY;
-        ps = Double.POSITIVE_INFINITY;
-        pc = Double.NEGATIVE_INFINITY;
+        qs = Float.POSITIVE_INFINITY;
+        qc = Float.NEGATIVE_INFINITY;
+        ps = Float.POSITIVE_INFINITY;
+        pc = Float.NEGATIVE_INFINITY;
 
 
         for (int y = 0; y < height; y++) {
